+++ server/API/api_tagRandum.js
@@ -1,13 +1,11 @@
 //======================================================================
 // set up
 //======================================================================
-const dbConf = require('../DbManager');
-const dbQuery = require('../lib/db_query');
-const df = require('../config/define');
+const apiFunc = require('../lib/api_functions');
 //======================================================================
 exports.getAPI = function (req,res) {
     let tag_num = req.params.num;
-    let results = dbQuery.getRandomArr(tag_num);
+    let results = apiFunc.getRandomArr(tag_num);
     return res.json({
         randum:results
     })
