+++ server/lib/db_query.js
@@ -1,6 +1,11 @@
-// db_query_functions
-
-// 豆知識10件取得
+/*=================================================================================
+ DB操作に関する関数集
+/*=================================================================================
+/**
+ * @description 豆知識を取得するSQL文を生成
+ * @param {Int} limit_n | 取得する豆知識の数
+ * @returns {String} sql | 豆知識取得SQL文
+ */
 exports.getTrivia = (limit_n) => {
     let sql = 'SELECT * FROM trivia_table AS tr ';
         sql += 'LEFT JOIN p_lang_mst AS pm ';
@@ -10,7 +15,12 @@ exports.getTrivia = (limit_n) => {
 
     return sql;
 }
-// 豆知識登録
+/**
+ * @description 豆知識を登録するSQL文を生成する
+ * @param {Int} p_lang_id | プログラミング言語のカラー番号
+ * @param {Int} article | 豆知識投稿内容
+ * @returns {String} sql | 豆知識登録SQL文
+ */
 exports.insTrivia = (p_lang_id, article) => {
     let sql = 'INSERT INTO trivia_table SET ';
         sql += 'p_lang_id = ' + p_lang_id;
@@ -21,44 +31,26 @@ exports.insTrivia = (p_lang_id, article) => {
 
     return sql;
 }
-// PG言語色取得 => 言語IDに対して
+/**
+ * @description PG言語色取得 => 言語IDに対して
+ * @param {Int} p_lang_id | プログラミング言語のカラー番号
+ * @returns {String} sql | プログラミング言語のカラー取得SQL文
+ */
 exports.getSPcolor = (p_lang_id) => {
     let sql = 'SELECT * FROM p_lang_mst ';
         sql += 'WHERE p_lang_id = ' + p_lang_id;
 
     return sql;
 }
-
-//  trivia_tableから最近投稿された言語のp_lang_idを10個取得する
-exports.getRecentlyLang = (n) => {
+/**
+ * @description trivia_tableから最近投稿された言語のp_lang_idを10個取得する
+ * @param {Int} limit_n | 取得する豆知識の数
+ * @returns {String} sql | 豆知識取得SQL文
+ */
+exports.getRecentlyLang = (limit_n) => {
     let sql = 'SELECT p_lang_id FROM ';
         sql += 'trivia_table GROUP BY p_lang_id ';
-        sql += 'ORDER BY MAX(ins_dt) DESC limit ' + n;
+        sql += 'ORDER BY MAX(ins_dt) DESC limit ' + limit_n;
 
     return sql;
-}
-
-//重複無しランダム配列
-exports.getRandomArr = (maxnum) => {
-    /** 重複チェック用配列 */
-    var randoms = [];
-    /** 最小値と最大値 */
-    var min = 0, max = maxnum-1;
-    
-    /** 重複チェックしながら乱数作成 */
-    for(i = min; i <= max; i++){
-        while(true){
-            var tmp = intRandom(min, max);
-            if(!randoms.includes(tmp)){
-            randoms.push(tmp);
-            break;
-            }
-        }
-    }
-    return randoms;
-    
-    /** min以上max以下の整数値の乱数を返す */
-    function intRandom(min, max){
-    return Math.floor( Math.random() * (max - min + 1)) + min;
-    }
 }
